Add Logger.detached factory
Sometimes you just don't need to add a logger, which will be a singleton
and part of the global loggers tree. So, just a logger, which will be
garbage collected, and its onRecord.listen won't subscribe to the root one.
Right now, there is theoretical possibility for that, but a constructor
for that is private.

Added Logger.detached factory, which is doing exactly this.