Replace _rootLevel with root._level
Previously root._level and _rootLevel both existed and could be
different.  The existing code took care to always use _rootLevel
instead of root._level, but I think it'd be better to not have the
extra Level variable at all.

This also provides an opportunity to expose a public constant for the
the default logging level, further addressing
https://github.com/dart-lang/logging/issues/57.