Override empty stack traces for traced events
We already default to the current stack trace of the logging caller when
there is no stack trace available so the author has some signal for
where to look in the code.

Also default for non-null but empty stack traces. This is more useful
where an empty trace was passed through some signature to satisfy
nullability requirements and forwarded to the log.